Refactor Slice prop Imports
We were previously spreading the slice props in SliceZone.astro,
which messed with typing.
We were also destructuring all of the props as they came into the slice.
Now we pass slice={slice}, type the props, and keep the context through
the component with `{slice.primary.heading}`

Credit to Angelo Ashmore for suggesting the refactor